Add test suite to determine when max image size does not support text
Add initial TextSetting#set test suite to detect when the text requires an image width larger than 3840(4K) pixels.

Rubocop is left unpleased in this commit because the functionality that computes the image related data such as width and height will probably be moved elsewhere, as the computation of this data is not a responsibility of the TextSetting class.