Add more tests to ensure helpful error messages for TextSetting#set
Add more tests to ensure that the errors raised by TextSetting#set contain helpful and contextual messages that hint the user how the error can be avoided/resolved.

Rubocop is still left unpleased in this commit because the functionality that computes the image related data such as width and height will probably be moved elsewhere, as the computation of this data is not a responsibility of the TextSetting class.

This change will probably be implemented once the Renderer class is implemented.