Finalize test suite to ensure secure Setting setters and getters
Add last tests to ensure that TextSetting setters and getters work as expected, no state is leaked and no state can be injected directly.

Rubocop is still unpleased for the TextSetting and Result classes because the Renderer will be implemented soon, which will hopefully result in a DRYer implementation and automatically get rid of Rubocop issues.

Faking implementation details to please Rubocop, at this point, is a waste of time.