Add more AsciiPngfy::Result tests
The Result#settings object was supposed to act exactly the same as the setting interface of the Pngfy instance, this is why the same exact ConfigurableSettings/SetableGetableSettings instance seemed like the perfect fit, both for keeping track of internal settings and to inform the user of the internal setting state through the same interface, but allowing only getters in the context of the result.

No new logic was needed for this approach since the ConfigurableSettings class uses a hash to determine wether specific setter and getter calls should be forwarded to the Setting itself or not.

This hash, which represents wether getters and/or setter calls are allowed is changed so that setters are not supported once it becomes an attribute of the Result.

The problem is the awkward hierarchy of Settings classes. The SetableGetableSettings is still used only as an easy way to setup a ConfigurableSettings instance with the expected setable and getable options, but since the SetableGetableSettings instance has the setter functionality disabled in the context of the Result, the object type SetableGetableSettings for the Result#settings is misleading.

To remedy this I will either add a GetableSettings class to clarify the intention of the Result#settings or remove the SetableGetableSettings class entirely and use the ConfigurableSettings for internal and Result settings.