Fix AsciiPngfy::Result#settings test suite
The AsciiPngfy::Result#settings tests were flawed in that the Result#settings data to be tested was pulled from the according result object right after the result was created, which was not the feature to be tested.

The fix required changing one setting, immediately generating a result object for that change, then change another setting and immediately generate a result object for that second change, then pull the data from both result objects and make sure that the oldest result object does not reference the same data as the most recent result object.

This is important because the result objects are coupled with the generated image along with other data points. If any of these data points gets mixed up, the result object does not make sense as a snapshot anymore.