Finalize initial Results and Glyphs test suite and generally refactor
Break up the nested classes AABB and Vec2i into their separate files at the AsciiPngfy Gem top-level namespace AsciiPngfy::AABB and AsciiPngfy::Vec2i.

Please and/or disable rubocop rules for the whole implementation apart from the TextSetting class, which still computes result related data such as the resulting png image dimensions. This was not yet flushed out by the test suites because the computations are the exact same. The TextSetting is the next thing to be re-designed and refactored so that there is only a single implementation of these computations, which are, if needed, used elsewhere.

At this point the core functionality is implemented but the implementation needs massive cleanup, especially the tests, which are integration heavy, need unit testing and re-structuring to at least partially concretize the parts of the achitecture that benefit from it and that do not change any time soon.