Architecture changes regarding the Renderer and TextSetting refactor
Renamed the renderer.rb file to settings_renderer.rb because that is the name of the class in that file.

Add RenderingRules module to lay open the logic to compute rendering related data. This was mostly done so that the TextSetting can pre-compute the image dimensions based on the text and the previously defined settings in order to determine wether the desired text is valid in terms of the max image resolution or not.

Future changes probably include a re-design of the Settings related functionality, because the dynamic nature of the ConfigurableSettings is really not needed, and could be replaced with something that is simpler, involves less deep object duplication, i.e. something that is more appropriate.

Future changes also probably include the way each Setting implementation is added to the Setting. The Setting implementations are currently added to a hash line by line, because the Hash notation along with the keyword 'self' causes some issues of the object its represents.

Currently the Settings objects is passed to the TextSetting that it itself contains, as the TextSetting must be able to access these settings to validate itself. This might also change in the future as everything seems coupled more than it needs to be.