Fix zero division and NAN error in color alpha component calculation
This division by zero and/or computation resulting in NAN was the result of setting the font_color and the background_color settings alpha values to zero.

This case slipped through because the settings have never been tested for the case when both the font_color and the background_color were exactly zero.

I will soon look into the correct computational procedure to calculate the color and color alpha values based on 'straight alpha composition' because the way it is currently done contains rounding error and makes too many assumptions about the nature of the numbers involved.

The solution will maybe even entail a change from straight alpha composition to pre-multiplied alpha composition as this procedure is much more precise in terms of color generation/representation and more numerically robust.