Simplify the Settings and the Settings snapshot implementation
The configurable Settings were always used for both setters and getters, so there was no advantage to having configurable logic. The original idea was to 'disable' the Settings setters for the snapshot, but this was unnecessarily complex because of the duplication involved and as mentioned before, never actually used.

In the new implementation, the Settings are setable and getable by default and not configurable anymore. The Settings object can now generate a SettingsSnapshot of itself which is just a dynamic wrapper for the setting names associated with their setting value.