Remove require of rubygems and don't enforce bundler programmatically
The Gem should not enforce Bundler programmatically in order to not force the behaviour onto the Gem user and because the user might use the Gemfile to manage the depencies of that project the Gem is used in.

But because the 'bundle', i.e. the Gemfile/Gemfile.lock must be enforced in terms of this Ruby project, invoking the interpreter must follow the 'bundler exec' call in order to add the resources specified by the Gemfile/Gemfile.lock and gemspec to the interpreter load path.