Change test file require types and update Rakefile
Test file 'require's changed from relative_require to require for two reasons. First it is easier to maintain, despite the fact that this project does not really need this for easy maintenance sake and second because it seems like the generally preferred/conventional method.

This requires a workaround for specific situations though, in this case, the testing_prerequisites file must be required by all the test files, which is fine for relative_require because the test files can just indicate the correct relative directory, but that does not work when using require instead, because the /test directory is not added to the , which means that the test files cannot require the test_prerequisites file.

In this case the problem is solved at the Rakefile level. To run all the tests in the /test directory, each test_*.rb file in /test and test sub-directory is iterated and interpreted one by one, with the catch that evertime the Ruby interpreter is invoked, the /test directory is added to the  using the -I flag.

This ensures that for each interpreted test file, the necessary files in the test directory can be loaded using 'require'.