Add initial documentational and portfolio-centric README.md content
This initial content is more of a brainstorming in order to determine what sort of content the README.md should contain and how it should look.

While the Portfolio aspect is the most important to me in terms of the creation of this gem, I still want this Gem to grow into a well designed and performant Gem that is actually used for what it was intended to. Therefore, the README.md must fulfill the documentational part and possibly refer to other sources that provide the Gem user with all the knowledge needed to fully leverage the functionality of this Gem.